Исправила ошибки, связанные со стилем java кода Удалила неиспользуемый код Random создается один раз Исправила getTwoCats (лишнее условие и возможнсть бесконечного цикла) Удалила лишний try-catch в isValid Удалила роли Картинки загружаются из uploads
