+++ src/main/java/com/acasssta/chooseCats/controllers/RegistrationController.java
@@ -40,14 +40,12 @@ public class RegistrationController {
         return "registration-form";
     }
 
-    // Binding Result после @ValidModel !!!
     @PostMapping("/processRegistrationForm")
     public String processRegistrationForm(@Valid @ModelAttribute("systemUser") SystemUser theSystemUser, Model theModel) {
         String userName = theSystemUser.getUserName();
 
         User existing = userService.findByUserName(userName);
         if (existing != null) {
-            // theSystemUser.setUserName(null);
             theModel.addAttribute("systemUser", theSystemUser);
             theModel.addAttribute("registrationError", "User with current username already exists");
             logger.debug("User name already exists.");
