--- src/main/java/com/acasssta/chooseCats/entities/Role.java
@@ -1,36 +1,0 @@
-package com.acasssta.chooseCats.entities;
-
-import lombok.Data;
-import org.springframework.security.core.GrantedAuthority;
-
-import javax.persistence.*;
-
-@Entity
-@Data
-@Table(name = "roles")
-public class Role implements GrantedAuthority {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "id")
-	private Long id;
-
-	@Column(name = "name")
-	private String name;
-
-	public Role() {
-	}
-
-	public Role(String name) {
-		this.name = name;
-	}
-
-	@Override
-	public String toString() {
-		return "Role{" + "id=" + id + ", name='" + name + '\'' + '}';
-	}
-
-	@Override
-	public String getAuthority() {
-		return name;
-	}
-}
