--- src/main/java/com/acasssta/chooseCats/repositories/RoleRepository.java
@@ -1,10 +1,0 @@
-package com.acasssta.chooseCats.repositories;
-
-import com.acasssta.chooseCats.entities.Role;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepository extends CrudRepository<Role, Long> {
-    Role findOneByName(String theRoleName);
-}
