+++ src/main/java/com/acasssta/chooseCats/services/UserServiceImpl.java
@@ -2,7 +2,6 @@ package com.acasssta.chooseCats.services;
 
 import com.acasssta.chooseCats.entities.*;
 import com.acasssta.chooseCats.repositories.CatRepository;
-import com.acasssta.chooseCats.repositories.RoleRepository;
 import com.acasssta.chooseCats.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.GrantedAuthority;
@@ -13,23 +12,13 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.Collectors;
+import java.util.*;
 
 @Service
 public class UserServiceImpl implements UserService {
     private UserRepository userRepository;
     private BCryptPasswordEncoder passwordEncoder;
     private CatRepository catRepository;
-    private RoleRepository roleRepository;
-
-    @Autowired
-    public void setRoleRepository(RoleRepository roleRepository) {
-        this.roleRepository = roleRepository;
-    }
 
     @Autowired
     public void setUserRepository(UserRepository userRepository) {
@@ -46,11 +35,6 @@ public class UserServiceImpl implements UserService {
         this.passwordEncoder = passwordEncoder;
     }
 
-    //    @Autowired
-//    public void setPasswordEncoder(BCryptPasswordEncoder passwordEncoder) {
-//        this.passwordEncoder = passwordEncoder;
-//    }
-
     @Override
     @Transactional
     public User findByUserName(String username) {
@@ -66,12 +50,10 @@ public class UserServiceImpl implements UserService {
             return false;
         }
 
+        user.setRole("user");
         user.setUserName(systemUser.getUserName());
         user.setPassword(passwordEncoder.encode(systemUser.getPassword()));
-
         user.setUserCat((List<Cat>) catRepository.findAll());
-        user.setRoles(Arrays.asList(roleRepository.findOneByName("ROLE_EMPLOYEE")));
-
         userRepository.save(user);
         return true;
     }
@@ -83,6 +65,11 @@ public class UserServiceImpl implements UserService {
         if (user == null) {
             throw new UsernameNotFoundException("Invalid username or password.");
         }
-        return new UserPrincipal(user);
+
+        Set<GrantedAuthority> authorities = new HashSet<GrantedAuthority>();
+        authorities.add(new SimpleGrantedAuthority(user.getRole()));
+
+        return new org.springframework.security.core.userdetails.User(user.getUserName(), user.getPassword(),
+                authorities);
     }
 }
