--- src/main/resources/db/choose_cat_bd/users_roles_540.sdi
@@ -1,1 +1,0 @@
-{"mysqld_version_id":80015,"dd_version":80014,"sdi_version":1,"dd_object_type":"Table","dd_object":{"name":"users_roles","mysql_version_id":80015,"created":20190310122823,"last_altered":20190310122823,"hidden":1,"options":"avg_row_length=0;key_block_size=0;keys_disabled=0;pack_record=0;stats_auto_recalc=0;stats_sample_pages=0;","columns":[{"name":"user_id","type":9,"is_nullable":false,"is_zerofill":false,"is_unsigned":false,"is_auto_increment":false,"is_virtual":false,"hidden":1,"ordinal_position":1,"char_length":20,"numeric_precision":19,"numeric_scale":0,"numeric_scale_null":false,"datetime_precision":0,"datetime_precision_null":1,"has_no_default":true,"default_value_null":false,"srs_id_null":true,"srs_id":0,"default_value":"AAAAAAAAAAA=","default_value_utf8_null":true,"default_value_utf8":"","default_option":"","update_option":"","comment":"","generation_expression":"","generation_expression_utf8":"","options":"interval_count=0;","se_private_data":"","column_key":4,"column_type_utf8":"bigint(20)","elements":[],"collation_id":8,"is_explicit_collation":false},{"name":"role_id","type":9,"is_nullable":false,"is_zerofill":false,"is_unsigned":false,"is_auto_increment":false,"is_virtual":false,"hidden":1,"ordinal_position":2,"char_length":20,"numeric_precision":19,"numeric_scale":0,"numeric_scale_null":false,"datetime_precision":0,"datetime_precision_null":1,"has_no_default":true,"default_value_null":false,"srs_id_null":true,"srs_id":0,"default_value":"AAAAAAAAAAA=","default_value_utf8_null":true,"default_value_utf8":"","default_option":"","update_option":"","comment":"","generation_expression":"","generation_expression_utf8":"","options":"interval_count=0;","se_private_data":"","column_key":4,"column_type_utf8":"bigint(20)","elements":[],"collation_id":8,"is_explicit_collation":false}],"schema_ref":"choose_cat_bd","se_private_id":18446744073709551615,"engine":"MyISAM","last_checked_for_upgrade_version_id":0,"comment":"","se_private_data":"","row_format":1,"partition_type":0,"partition_expression":"","partition_expression_utf8":"","default_partitioning":0,"subpartition_type":0,"subpartition_expression":"","subpartition_expression_utf8":"","default_subpartitioning":0,"indexes":[{"name":"FKj6m8fwv7oqv74fcehir1a9ffy","hidden":false,"is_generated":true,"ordinal_position":1,"comment":"","options":"flags=0;","se_private_data":"","type":3,"algorithm":2,"is_algorithm_explicit":false,"is_visible":true,"engine":"MyISAM","elements":[{"ordinal_position":1,"length":8,"order":2,"column_opx":1}]},{"name":"FK2o0jvgh89lemvvo17cbqvdxaa","hidden":false,"is_generated":true,"ordinal_position":2,"comment":"","options":"flags=0;","se_private_data":"","type":3,"algorithm":2,"is_algorithm_explicit":false,"is_visible":true,"engine":"MyISAM","elements":[{"ordinal_position":1,"length":8,"order":2,"column_opx":0}]}],"foreign_keys":[],"partitions":[],"collation_id":33}}
